Core (stats): k in percentile should be an int

+++ brevitas/core/stats.py
@@ -133,14 +133,14 @@ class AbsPercentile(brevitas.jit.ScriptModule):
     def forward(self, x):
         if self.stats_reduce_dim is None:
             # k is 1-indexed, so round away from zero
-            k = math.floor(.01 * self.q * x.numel() + 0.5)
+            k = int(math.floor(.01 * self.q * x.numel() + 0.5))
             result = x.abs().view(-1).kthvalue(k).values
         else:
             # assuming x is two dimensional, get the other dimension
             other_dim = abs(self.stats_reduce_dim - 1)
             dim_slice = torch.narrow(x, dim=other_dim, start=0, length=1)
             # k is 1-indexed, so round away from zero
-            k = math.floor(.01 * self.q * dim_slice.numel() + 0.5)
+            k = int(math.floor(.01 * self.q * dim_slice.numel() + 0.5))
             result = x.abs().kthvalue(k, dim=self.stats_reduce_dim).values
         return result
